Сolor generation moved to Generator, ColorConverter implemented
