+++ Right on target/Model/Generator.swift
@@ -5,15 +5,18 @@
 //  Created by Aleksei Frolov on 22.08.2024.
 //
 
-import Foundation
+import UIKit
 
 protocol GeneratorProtocol {
-    func getRandomValue() -> Int
+    func getRandomValue() -> Int?
+    func getRandomColor() -> UIColor
 }
 
 class Generator: GeneratorProtocol {
-    private var minBorderValue: Int
-    private var maxBorderValue: Int
+    private var minBorderValue: Int?
+    private var maxBorderValue: Int?
+    
+    init() {}
     
     init?(startValue: Int, endValue: Int) {
         guard startValue <= endValue else {
@@ -23,7 +26,16 @@ class Generator: GeneratorProtocol {
         maxBorderValue = endValue
     }
     
-    func getRandomValue() -> Int {
-        (minBorderValue...maxBorderValue).randomElement()!
+    func getRandomValue() -> Int? {
+        guard let minBorderValue, let maxBorderValue else { return nil }
+        return (minBorderValue...maxBorderValue).randomElement()
+    }
+    
+    func getRandomColor() -> UIColor {
+        UIColor(
+            red: CGFloat.random(in: 0...1),
+            green: CGFloat.random(in: 0...1),
+            blue: CGFloat.random(in: 0...1),
+            alpha: 1.0)
     }
 }
